requirements.txt
pytesseract
opencv-python
Pillow
lxml
